feat!: remove the dotMemory path and the ensure_dotmemory tool
BREAKING CHANGE: the ensure_dotmemory MCP tool is removed. Six tools become
five. Clients referencing it by name will get an unknown-tool error.

dotMemory Console cannot analyze anything -- it has no report command, and .dmw
workspaces need the paid standalone GUI. So snapshot has always written an
artifact analyze could never read. With collection now in-process, the installer,
the tool manager, the report parser and the analyzer are all unreferenced.

This also deletes the #118 execute-bit guards added earlier today. That is
deliberate: deleting DotMemoryAutoInstaller deletes the bug class those tests
guarded, which is stronger than guarding it.
