+++ src/MemoryLens.Mcp/Properties/AssemblyInfo.cs
@@ -1,4 +1,3 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("MemoryLens.Mcp.Tests")]
-[assembly: InternalsVisibleTo("MemoryLens.Mcp.IntegrationTests")]
