+++ tests/MemoryLens.Mcp.IntegrationTests/MemoryLens.Mcp.IntegrationTests.csproj
@@ -5,7 +5,7 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <IsPackable>false</IsPackable>
-    <TestingPlatformCommandLineArguments>--minimum-expected-tests 13</TestingPlatformCommandLineArguments>
+    <TestingPlatformCommandLineArguments>--minimum-expected-tests 10</TestingPlatformCommandLineArguments>
   </PropertyGroup>
 
   <ItemGroup>
