--- tests/MemoryLens.Mcp.Tests/Profiler/SnapshotManagerTests.cs
@@ -1,78 +1,0 @@
-using MemoryLens.Mcp.Profiler;
-using Xunit;
-
-namespace MemoryLens.Mcp.Tests.Profiler;
-
-public class SnapshotManagerTests
-{
-    [Fact]
-    public async Task TakeSnapshot_ByPid_ReturnsSnapshotId()
-    {
-        var runner = new FakeProcessRunner(exitCode: 0, output: "Snapshot saved");
-        var filter = new ProcessFilter();
-        var toolManager = new FakeDotMemoryToolManager(runner);
-        var manager = new SnapshotManager(runner, filter, toolManager);
-
-        var result = await manager.TakeSnapshotAsync(pid: 1234, ct: TestContext.Current.CancellationToken);
-
-        Assert.True(result.Success);
-        Assert.NotNull(result.SnapshotId);
-        Assert.Null(result.Error);
-    }
-
-    [Fact]
-    public async Task TakeSnapshot_ExcludedProcess_ReturnsError()
-    {
-        var runner = new FakeProcessRunner(exitCode: 0, output: "");
-        var filter = new ProcessFilter();
-        var toolManager = new FakeDotMemoryToolManager(runner);
-        var manager = new SnapshotManager(runner, filter, toolManager);
-
-        var result = await manager.TakeSnapshotAsync(processName: "devenv", ct: TestContext.Current.CancellationToken);
-
-        Assert.False(result.Success);
-        Assert.Contains("excluded", result.Error!);
-    }
-
-    [Fact]
-    public async Task TakeSnapshot_ByProcessName_ReturnsSnapshotId()
-    {
-        var runner = new FakeProcessRunner(exitCode: 0, output: "Snapshot saved");
-        var filter = new ProcessFilter();
-        var toolManager = new FakeDotMemoryToolManager(runner);
-        var manager = new SnapshotManager(runner, filter, toolManager);
-
-        var result = await manager.TakeSnapshotAsync(processName: "MyApp", ct: TestContext.Current.CancellationToken);
-
-        Assert.True(result.Success);
-        Assert.NotNull(result.SnapshotId);
-    }
-
-    [Fact]
-    public async Task TakeSnapshot_WithCommand_ReturnsSnapshotId()
-    {
-        var runner = new FakeProcessRunner(exitCode: 0, output: "Snapshot saved");
-        var filter = new ProcessFilter();
-        var toolManager = new FakeDotMemoryToolManager(runner);
-        var manager = new SnapshotManager(runner, filter, toolManager);
-
-        var result = await manager.TakeSnapshotAsync(command: "dotnet run --project MyApp", ct: TestContext.Current.CancellationToken);
-
-        Assert.True(result.Success);
-        Assert.NotNull(result.SnapshotId);
-    }
-
-    [Fact]
-    public async Task TakeSnapshot_ProcessRunnerFails_ReturnsError()
-    {
-        var runner = new FakeProcessRunner(exitCode: 1, output: "");
-        var filter = new ProcessFilter();
-        var toolManager = new FakeDotMemoryToolManager(runner);
-        var manager = new SnapshotManager(runner, filter, toolManager);
-
-        var result = await manager.TakeSnapshotAsync(pid: 1234, ct: TestContext.Current.CancellationToken);
-
-        Assert.False(result.Success);
-        Assert.Contains("dotMemory CLI failed", result.Error!);
-    }
-}
